requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
line-bot-sdk==3.16.1
httpx==0.28.1
aiohttp==3.11.11
